+++ tests/serializers.py
@@ -1,5 +1,5 @@
+from rest_framework_json_api import serializers
 from rest_framework_json_api.relations import ResourceRelatedField
-from rest_framework_json_api.serializers import ModelSerializer
 from tests.models import (
     BasicModel,
     ForeignKeySource,
@@ -9,13 +9,13 @@ from tests.models import (
 )
 
 
-class BasicModelSerializer(ModelSerializer):
+class BasicModelSerializer(serializers.ModelSerializer):
     class Meta:
         fields = ("text",)
         model = BasicModel
 
 
-class ForeignKeySourceSerializer(ModelSerializer):
+class ForeignKeySourceSerializer(serializers.ModelSerializer):
     target = ResourceRelatedField(queryset=ForeignKeyTarget.objects)
 
     class Meta:
@@ -23,7 +23,7 @@ class ForeignKeySourceSerializer(ModelSerializer):
         fields = ("target",)
 
 
-class ManyToManySourceSerializer(ModelSerializer):
+class ManyToManySourceSerializer(serializers.ModelSerializer):
     targets = ResourceRelatedField(many=True, queryset=ManyToManyTarget.objects)
 
     class Meta:
@@ -31,14 +31,21 @@ class ManyToManySourceSerializer(ModelSerializer):
         fields = ("targets",)
 
 
-class ManyToManyTargetSerializer(ModelSerializer):
+class ManyToManyTargetSerializer(serializers.ModelSerializer):
     class Meta:
         model = ManyToManyTarget
 
 
-class ManyToManySourceReadOnlySerializer(ModelSerializer):
+class ManyToManySourceReadOnlySerializer(serializers.ModelSerializer):
     targets = ResourceRelatedField(many=True, read_only=True)
 
     class Meta:
         model = ManyToManySource
         fields = ("targets",)
+
+
+class CallableDefaultSerializer(serializers.Serializer):
+    field = serializers.CharField(default=serializers.CreateOnlyDefault("default"))
+
+    class Meta:
+        fields = ("field",)